+++ src/io.rs
@@ -47,7 +47,7 @@ use tokio::{
 };
 
 use crate::{
-    header::Header,
+    header::{ErrorKind, Header},
     protocol::{
         payload_is_multi_frame, CompletedRead, FrameIter, JulietProtocol, LocalProtocolViolation,
         OutgoingFrame, OutgoingMessage, ProtocolBuilder,
@@ -200,15 +200,26 @@ pub enum CoreError {
         data.as_ref().map(|b| b.len()).unwrap_or(0))
     ]
     RemoteReportedError {
-        /// Header of the reported error.
+        /// Header of the received error.
         header: Header,
         /// The error payload, if the error kind was
         /// [`ErrorKind::Other`](crate::header::ErrorKind::Other).
         data: Option<Bytes>,
     },
     /// The remote peer violated the protocol and has been sent an error.
-    #[error("error sent to peer: {0}")]
-    RemoteProtocolViolation(Header),
+    #[error("error sent to peer [channel {}/id {}]: {} (payload: {} bytes)",
+        header.channel(),
+        header.id(),
+        header.error_kind(),
+        data.as_ref().map(|b| b.len()).unwrap_or(0))
+    ]
+    RemoteProtocolViolation {
+        /// Header of the locally created error.
+        header: Header,
+        /// The error payload, if the error kind was
+        /// [`ErrorKind::Other`](crate::header::ErrorKind::Other).
+        data: Option<Bytes>,
+    },
     #[error("local protocol violation")]
     /// Local protocol violation - caller violated the crate's API.
     LocalProtocolViolation(#[from] LocalProtocolViolation),
@@ -480,22 +491,36 @@ where
     /// Polling of this function must continue only until `Err(_)` or `Ok(None)` is returned,
     /// indicating that the connection should be closed or has been closed.
     pub async fn next_event(&mut self) -> Result<Option<IoEvent>, CoreError> {
-        if let Some(ref mut pending_error) = self.pending_error {
-            tokio::time::timeout(self.error_timeout, self.writer.write_all_buf(pending_error))
-                .await
-                .map_err(|_elapsed| CoreError::ErrorWriteTimeout)?
-                .map_err(CoreError::WriteFailed)?;
-
-            // We succeeded writing, clear the error.
-            let peers_crime = self
-                .pending_error
-                .take()
-                .expect("pending_error should not have disappeared")
-                .header();
-            return Err(CoreError::RemoteProtocolViolation(peers_crime));
-        }
-
         loop {
+            if let Some(ref mut pending_error) = self.pending_error {
+                // We have to invert the process of turning an `Option<Bytes>` into a `Bytes` here,
+                // since sending it will consume the bytes in the buffer. Rely on the fact that
+                // there is a payload if and only if the error kind is OTHER.
+                let data = if pending_error.header().is_error()
+                    && pending_error.header().error_kind() == ErrorKind::Other
+                {
+                    Some(pending_error.segment().clone())
+                } else {
+                    None
+                };
+
+                tokio::time::timeout(self.error_timeout, self.writer.write_all_buf(pending_error))
+                    .await
+                    .map_err(|_elapsed| CoreError::ErrorWriteTimeout)?
+                    .map_err(CoreError::WriteFailed)?;
+
+                // We succeeded writing, clear the error.
+                let peers_crime = self
+                    .pending_error
+                    .take()
+                    .expect("pending_error should not have disappeared");
+
+                return Err(CoreError::RemoteProtocolViolation {
+                    header: peers_crime.header(),
+                    data,
+                });
+            }
+
             if self.next_parse_at <= self.buffer.remaining() {
                 // Simplify reasoning about this code.
                 self.next_parse_at = 0;
@@ -508,7 +533,7 @@ where
                     Outcome::Fatal(err_msg) => {
                         // The remote messed up, begin shutting down due to an error.
                         #[cfg(feature = "tracing")]
-                        tracing::warn!(err_msg_header=%err_msg.header(), "injecting error due to fatal outcome");
+                        tracing::debug!(err_msg_header=%err_msg.header(), "injecting error due to fatal outcome");
                         self.inject_error(err_msg);
                     }
                     Outcome::Success(successful_read) => {
@@ -571,7 +596,10 @@ where
 
                         if header_sent.is_error() {
                             // We finished sending an error frame, time to exit.
-                            return Err(CoreError::RemoteProtocolViolation(header_sent));
+                            return Err(CoreError::RemoteProtocolViolation{
+                                header: header_sent,
+                                data: None
+                            });
                         }
 
                         // TODO: We should restrict the dirty-queue processing here a little bit
@@ -850,6 +878,36 @@ where
         }
 
         Ok(())
+    }
+}
+
+impl CoreError {
+    /// If the error stems from a peer *sending* us a custom error, return its header and payload.
+    #[inline(always)]
+    pub fn as_remote_other_err(&self) -> Option<(Header, &Bytes)> {
+        match self {
+            CoreError::RemoteReportedError { header, data }
+                if header.is_error() && header.error_kind() == ErrorKind::Other =>
+            {
+                debug_assert!(data.is_some(), "`CoreError::RemoteReportedError` should never have no `data` for `OTHER` kind of error");
+                Some((*header, data.as_ref()?))
+            }
+            _ => None,
+        }
+    }
+
+    /// If the error was a locally created custom error, return its header and payload.
+    #[inline(always)]
+    pub fn as_local_other_err(&self) -> Option<(Header, &Bytes)> {
+        match self {
+            CoreError::RemoteProtocolViolation { header, data }
+                if header.is_error() && header.error_kind() == ErrorKind::Other =>
+            {
+                debug_assert!(data.is_some(), "`CoreError::RemoteReportedError` should never have no `data` for `OTHER` kind of error");
+                Some((*header, data.as_ref()?))
+            }
+            _ => None,
+        }
     }
 }